+++ apps/api/views/qrcode_views.py
@@ -1,3 +1,5 @@
+import logging
+
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
@@ -11,8 +13,10 @@ from django.db import IntegrityError
 
 from ..serializers import QRCodeSerializer, AllQRCodesSerializer
 
-from ..services.qrcode_service import create_qrcode_by_branch_id_service, get_qrcode_by_id_service,\
+from ..services.qrcode_service import create_qrcode_by_branch_id_service, get_qrcode_by_id_service, \
     get_all_qrcode_service
+
+logger = logging.getLogger('django')
 
 
 @swagger_auto_schema(
@@ -70,25 +74,32 @@ def create_qrcode_view(request):
         )
         if qr_code:
             serializer = QRCodeSerializer(qr_code, many=False, context={"request": request})
+            message = 'Запрос выполнен успешно'
+            logger.info(f'{__name__} - {message} / user id:{user.id}')
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(data={'detail': 'Это не ваш branch'}, status=status.HTTP_403_FORBIDDEN)
+            message = 'Это не ваш branch'
+            logger.warning(f'{__name__} - {message} / user id:{user.id}')
+            return Response(data={'detail': message}, status=status.HTTP_403_FORBIDDEN)
 
     except ObjectDoesNotExist as e:
-        return Response(data={'detail': f'Такой branch не найден {e}'}, status=status.HTTP_404_NOT_FOUND)
+        message = f'Такой branch не найден {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except KeyError as e:
         message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     except IntegrityError as e:
-        return Response(
-            data={'detail': f'QRCode уже создан для этого branch {e}'},
-            status=status.HTTP_406_NOT_ACCEPTABLE
-        )
+        message = f'QRCode уже создан для этого branch {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        return Response(data={'detail': message}, status=status.HTTP_406_NOT_ACCEPTABLE)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
+        logger.critical(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -128,15 +139,22 @@ def read_qrcode_view(request, pk):
         qr_code = get_qrcode_by_id_service(user=user, qrcode_id=pk)
         if qr_code:
             serializer = QRCodeSerializer(qr_code, many=False, context={"request": request})
+            message = 'Запрос выполнен успешно'
+            logger.info(f'{__name__} - {message} / user id:{user.id}')
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         else:
-            return Response(data={'detail': 'Это не ваш branch'}, status=status.HTTP_403_FORBIDDEN)
+            message = 'Это не ваш branch'
+            logger.warning(f'{__name__} - {message} / user id:{user.id}')
+            return Response(data={'detail': message}, status=status.HTTP_403_FORBIDDEN)
 
     except ObjectDoesNotExist as e:
-        return Response(data={'detail': f'Такой QRCode не найден {e}'}, status=status.HTTP_404_NOT_FOUND)
+        message = f'Такой QRCode не найден {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
+        logger.critical(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -167,11 +185,16 @@ def read_all_qrcodes_view(request):
     try:
         qr_code_list = get_all_qrcode_service()
         serializer = AllQRCodesSerializer(qr_code_list, many=True, context={"request": request})
+        message = 'Запрос выполнен успешно'
+        logger.info(f'{__name__} - {message}')
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     except ObjectDoesNotExist as e:
-        return Response(data={'detail': f'Такой QRCode не найден {e}'}, status=status.HTTP_404_NOT_FOUND)
+        message = 'QRCodes не найдены'
+        logger.warning(f'{__name__} - {message}')
+        return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
+        logger.critical(f'{__name__} - {message}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)