+++ apps/api/views/rate_views.py
@@ -1,3 +1,5 @@
+import logging
+
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
@@ -12,6 +14,8 @@ from django.db import IntegrityError
 from ..serializers import RateSerializer
 
 from ..services.rate_service import get_rate_by_user_service, create_rate_by_user_service
+
+logger = logging.getLogger('django')
 
 
 @swagger_auto_schema(
@@ -63,17 +67,23 @@ def create_rate_view(request):
             user=user,
         )
         serializer = RateSerializer(rate, many=False)
+        message = 'Запрос выполнен успешно'
+        logger.info(f'{__name__} - {message} / user id:{user.id}')
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     except KeyError as e:
         message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     except IntegrityError as e:
-        return Response(data={'detail': f'Rate уже создан для этого user {e}'}, status=status.HTTP_406_NOT_ACCEPTABLE)
+        message = f'Rate уже создан для этого user {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        return Response(data={'detail': message}, status=status.HTTP_406_NOT_ACCEPTABLE)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
+        logger.critical(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -112,13 +122,18 @@ def read_rate_view(request):
     try:
         rate = get_rate_by_user_service(user=user)
         serializer = RateSerializer(rate, many=False)
+        message = 'Запрос выполнен успешно'
+        logger.info(f'{__name__} - {message} / user id:{user.id}')
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     except ObjectDoesNotExist as e:
-        return Response(data={'detail': f'Такой rate не найден {e}'}, status=status.HTTP_404_NOT_FOUND)
+        message = f'Такой rate не найден {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
+        logger.critical(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -171,15 +186,21 @@ def update_rate_view(request):
         serializer = RateSerializer(rate, many=False, partial=True, data=data)
         if serializer.is_valid():
             serializer.save()
+        message = 'Запрос выполнен успешно'
+        logger.info(f'{__name__} - {message} / user id:{user.id}')
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     except ObjectDoesNotExist as e:
-        return Response(data={'detail': f'Такой rate не найден {e}'}, status=status.HTTP_404_NOT_FOUND)
+        message = f'Такой rate не найден {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except KeyError as e:
         message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        logger.warning(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
+        logger.critical(f'{__name__} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)